Modificando models, forms, views, admin. Adicionando temaplates e static

+++ aluno/views.py
@@ -1,44 +1,53 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from .models import Aluno
-from .forms import AlunoForms
+from .forms import AlunoForm
 
+# Página inicial
 def index(request):
     alunos_total = Aluno.objects.count()
-    return render(request, 'aluno/index.html', {'alunos_total': alunos_total})
+    return render(request, 'aluno/home.html', {'alunos_total': alunos_total})
 
+# Listagem de alunos
 def alunos(request):
     alunos = Aluno.objects.all()
     return render(request, 'aluno/alunos.html', {'alunos': alunos})
 
-def alunos_detail(request):
-    alunos = Aluno.objects.all()
-    return render(request, 'aluno/alunos_detail.html', {'alunos': alunos})
+# Detalhes de um aluno
+def alunos_detail(request, pk):
+    aluno = get_object_or_404(Aluno, pk=pk)
+    return render(request, 'aluno/detalhes.html', {'aluno': aluno})
 
+# Criar aluno
 def alunos_create(request):
-    form = AlunoForms(request.POST or None)
     if request.method == 'POST':
+        form = AlunoForm(request.POST)
         if form.is_valid():
             form.save()
             return redirect('alunos')
     else:
-        form = AlunoForms()
-    return render(request, 'aluno/alunos_forms.html', {'form': form})
+        form = AlunoForm()
+    return render(request, 'aluno/aluno_forms.html', {'form': form})
 
+# Editar aluno
 def alunos_update(request, pk):
     aluno = get_object_or_404(Aluno, pk=pk)
-    form = AlunoForms(request.POST or None, instance=aluno)
     if request.method == 'POST':
+        form = AlunoForm(request.POST, instance=aluno)
         if form.is_valid():
-            form.save
+            form.save()  # <- aqui tinha faltado os parênteses
             return redirect('alunos')
     else:
-        form = AlunoForms(instance=aluno)
-    return render(request, 'aluno/alunos_forms.html', {'form': form})
+        form = AlunoForm(instance=aluno)
+    return render(request, 'aluno/aluno_forms.html', {'form': form})
 
+# Remover aluno
 def alunos_delete(request, pk):
     aluno = get_object_or_404(Aluno, pk=pk)
-    aluno.delete()
-    return redirect('alunos')
+    if request.method == 'POST':
+        aluno.delete()
+        return redirect('alunos')
+    return render(request, 'aluno/aluno_delete.html', {'aluno': aluno})
+
